Added (some) server side stuff & some upload stuff
Added log-in autherization. Right now there is only one account that can
log-in to the firebase database, the credentials are currently hardcoded
in so when app is run you can just click log-in instead of typing
credentials everytime, but the actual code to authenticate is actually
there, just commented out to make our lives easier.
Added functionality to log-out button.
Added functionality to choose from gallery button and take a picture
button in upload page, though i havent added code to story image to
database yet.